pyproject.toml
[tool.poetry]
name = "ruvon-sdk"
version = "0.1.0"
description = "Offline-first workflow engine for edge devices - saga compensation, store-and-forward, fleet config rollout"
authors = ["Ruvon Team <team@ruvon.dev>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/KamikaziD/ruvon-core"
repository = "https://github.com/KamikaziD/ruvon-core"
keywords = ["workflow", "fintech", "pos", "edge", "offline", "store-and-forward", "saga", "payment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial :: Point-Of-Sale",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [
    { include = "ruvon", from = "src" },
    { include = "ruvon_cli", from = "src" },
]

exclude = [
    "src/ruvon/examples",
    "src/ruvon/examples/**/*",
]

[tool.poetry.dependencies]
python = "^3.9"

# Core dependencies (required)
pydantic = "^2.0"
PyYAML = "^6.0"
jinja2 = "^3.1"
typer = ">=0.21"
click = "^8.0"
aiosqlite = ">=0.19"
anyio = "^4.0"
orjson = "^3.9"
msgspec = ">=0.18"
httpx = ">=0.25"
cryptography = ">=41.0"
sqlalchemy = "^2.0"
alembic = "^1.13"
python-dotenv = "^1.0"
typing-extensions = "^4.0"
annotated-types = ">=0.7.0"
jsonschema = "^4.17,<5.0"

# Workflow scheduling
croniter = ">=2.0"

# PostgreSQL support (optional)
asyncpg = {version = ">=0.27", optional = true}

# CLI and performance (optional)
rich = {version = ">=13.0", optional = true}
uvloop = {version = ">=0.19", optional = true}

# OpenTelemetry (optional)
opentelemetry-sdk = {version = ">=1.20", optional = true}
opentelemetry-exporter-otlp-proto-grpc = {version = ">=1.20", optional = true}

# NATS JetStream + Protobuf (optional)
nats-py = {version = ">=2.3", optional = true}
betterproto = {version = ">=0.9.0", optional = true}
protobuf = {version = ">=4.21", optional = true}

# AI Workflow Builder (optional) — Ollama uses httpx (already a core dep)
anthropic = {version = ">=0.25", optional = true}

# Knowledge base / RAG (optional)
lancedb = {version = ">=0.6", optional = true}
fastembed = {version = ">=0.3", optional = true}
rank-bm25 = {version = ">=0.2", optional = true}
psutil = {version = ">=5.9", optional = true}
sqlite-vec = {version = ">=0.1", optional = true}

[tool.poetry.extras]
postgres = ["asyncpg"]
cli = ["rich"]
performance = ["uvloop"]
otel = ["opentelemetry-sdk", "opentelemetry-exporter-otlp-proto-grpc"]
nats = ["nats-py", "betterproto", "protobuf"]
builder = ["anthropic"]
rag = ["lancedb", "fastembed", "rank-bm25", "psutil"]
rag-edge = ["fastembed", "sqlite-vec", "rank-bm25", "psutil"]
builder-full = ["anthropic", "lancedb", "fastembed", "rank-bm25", "psutil"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.21"
pytest-cov = "^4.0"
pytest-mock = "^3.12"
flake8 = "^6.0"
black = "^23.0"
mypy = "^1.0"

[tool.poetry.scripts]
ruvon = "ruvon_cli.main:app"

[build-system]
requires = ["poetry-core>=1.0.0,<2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/ruvon --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
